requirements.txt
flask
tweepy
gunicorn
python-multipart
python-dotenv
google-api-python-client
flask-cors